# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin restful",
        "title": "商城系统后台接口文档",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jimmy",
            "url": "http://www.swagger.io/support",
            "email": "tangzhiming90@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "8000",
    "basePath": "web/api/v1",
    "paths": {
        "/category/create": {
            "post": {
                "description": "需要传token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "web后台系统创建类目接口",
                "parameters": [
                    {
                        "description": "类目创建参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WebCategoryCreateParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"创建成功\", count, c}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "通过token校验通过后才可以登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "web后台系统登录接口",
                "parameters": [
                    {
                        "description": "用户登录参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WebLoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WebUserInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.WebCategoryCreateParam": {
            "type": "object",
            "required": [
                "level",
                "name",
                "parentId",
                "sort"
            ],
            "properties": {
                "level": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "models.WebLoginParam": {
            "type": "object",
            "properties": {
                "captchaId": {
                    "description": "验证码接口返回的字符串",
                    "type": "string"
                },
                "captchaValue": {
                    "description": "验证码",
                    "type": "string"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "username": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
        "models.WebUserInfo": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "uid": {
                    "type": "integer"
                }
            }
        }
    }
}